% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citenn.R
\name{citenn}
\alias{citenn}
\title{Creates a 95% confidence interval for treatment effects using neural networks}
\usage{
citenn(
  x,
  t,
  y,
  arcitecture_treatment = c(128, 64, 32),
  arcitecture_propensity = c(128, 64, 32),
  random = FALSE,
  train_proportion = 0.8,
  epochs_treatment = 30,
  epochs_propensity = 30
)
}
\arguments{
\item{x}{A n x m matrix where n is the number of observations and m is the number of explanatory variables excluding the treatment status.}

\item{t}{A n x 1 matrix containing the treatment status(1=treated,0=non treated)}

\item{y}{A n x 1 matrix containing the target variable}

\item{arcitecture_treatment}{A vector where each entry stands for a hidden layer and the value represents the number of hidden nodes in that layer.}

\item{arcitecture_propensity}{A vector where each entry stands for a hidden layer and the value represents the number of hidden nodes in that layer.}

\item{random}{A boolean indicating if treatment was randomised or not}

\item{train_proportion}{The proportion of data used for training}

\item{epochs_treatment}{The maximum number of epochs the treatment FFN is fitted for.}

\item{epochs_propensity}{The maximum number of epochs the propensity FFN is fitted for.}
}
\value{
returns a list containing: 1.Confidence interval lower bound 2.Confidence interval upper bound 3.Predictions ofmu0
4.Predictions of tau 5.Predictions of y 6.Predicted probability of treatment
}
\description{
Creates a 95% confidence interval for treatment effects using neural networks
}
